# lib/db/src/schema/devices.ts
import { pgTable, text, timestamp, uuid, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema, createSelectSchema } from "drizzle-zod";
import { z } from "zod";
import { usersTable } from "./users";

export const devicesTable = pgTable("devices", {
  id: uuid("id").primaryKey().defaultRandom(),
  userId: uuid("user_id")
    .notNull()
    .references(() => usersTable.id, { onDelete: "cascade" }),
  deviceName: text("device_name").notNull(),
  deviceModel: text("device_model"),
  androidVersion: text("android_version"),
  appVersion: text("app_version"),
  pushToken: text("push_token"),
  metadata: jsonb("metadata").$type<Record<string, unknown>>(),
  lastSeenAt: timestamp("last_seen_at", { withTimezone: true }).defaultNow().notNull(),
  createdAt: timestamp("created_at", { withTimezone: true }).defaultNow().notNull(),
});

export const insertDeviceSchema = createInsertSchema(devicesTable).omit({
  id: true,
  createdAt: true,
  lastSeenAt: true,
});

export const selectDeviceSchema = createSelectSchema(devicesTable);

export type InsertDevice = z.infer<typeof insertDeviceSchema>;
export type Device = typeof devicesTable.$inferSelect;